--- app/actions.py
@@ -1,5 +1,0 @@
-from server import db, app
-
-
-with app.app_context():
-    db.create_all()
