+++ boot.bat
@@ -1,0 +1,3 @@
+SET FLASK_APP=run.py
+SET FLASK_DEBUG=1
+flask run
