+++ run.py
@@ -1,0 +1,3 @@
+from app import create_app
+
+create_app("development")
